--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.2.13-dev1
+
+* **Optimize `MilvusUploadStager.parse_date_string`**
+
+## 1.2.13-dev0
+
+* **Optimize `parse_date_string`**
+
 ## 1.2.12
 
 * **Fix: retry with wait when throttling error happens in Sharepoint connector**