--- a/unstructured_ingest/__version__.py
+++ b/unstructured_ingest/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.13-dev1"  # pragma: no cover
+__version__ = "1.2.17-dev1"  # pragma: no cover